% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.lf.tgfb.mcti-package.R
\docType{package}
\name{rnaseq.lf.tgfb.mcti-package}
\alias{rnaseq.lf.tgfb.mcti}
\alias{rnaseq.lf.tgfb.mcti-package}
\title{rnaseq.lf.tgfb.mcti: RNA-seq Analysis of Lactate Transporter Inhibition in Myofibroblasts}
\description{
This package is a data repository for RNA-seq data obtained to study the impact of lactate transport inhibition on myofibroblast differentiation. Normal lung fibroblasts (Lonza) were treated with TGFβ (2 ng/mL) in combination with MCT1 inhibitor AZD3965 (100 nM) or MCT4 inhibitor VB124 (10 μM), separately or together, with DMSO (0.1%) as the vehicle control. RNA was obtained after 48 h of treatment and submitted for sequencing.
}
\author{
\strong{Maintainer}: William Oldham \email{wm.oldham@gmail.com} (\href{https://orcid.org/0000-0003-3029-4866}{ORCID})

}
\keyword{internal}
